requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.6.0
google-generativeai>=0.3.0
pytest>=7.0.0
pytest-asyncio>=0.19.0